+++ busy_beaver/models/__init__.py
@@ -2,9 +2,9 @@
 from .base import BaseModel  # noqa
 
 # Models
-from .user import GitHubSummaryUser  # noqa
 from .api_user import ApiUser  # noqa
 from .event import Event  # noqa
+from .github_summary_user import GitHubSummaryUser  # noqa
 from .key_value import KeyValueStore  # noqa
 from .slack import SlackInstallation  # noqa
 from .task import (  # noqa
